我的简历业务修改1-回收站返回漏洞

+++ src/main/java/com/zyh/easyapplyresume/service/impl/user/UserDeleteResumeServiceImpl.java
@@ -66,11 +66,13 @@ public class UserDeleteResumeServiceImpl implements UserDeleteResumeService {
             if (userDeleteResumes != null){
                 return BeanUtil.copyToList(userDeleteResumes, UserDeleteResumeInfoVO.class);
             }
+            return null;
+        }catch (BusException e){
+            throw e;
         }catch (Exception e){
-            log.error("查询用户删除简历信息失败！");
+            log.error("查询用户删除简历信息失败！", e);
+            throw new RuntimeException("查询用户删除简历信息失败");
         }
-
-        return null;
     }
 
     @Override
@@ -94,13 +96,6 @@ public class UserDeleteResumeServiceImpl implements UserDeleteResumeService {
             queryWrapper.eq(UserDeleteResume::getUserDeleteResumeUserId, userSaveResumeUserId);
             queryWrapper.orderByDesc(UserDeleteResume::getUserDeleteResumeSortedNum);
             List<UserDeleteResume> userDeleteResumes = userDeleteResumeMapper.selectList(queryWrapper);
-            Integer userDeleteResumeSortedNum = 0;
-            if (userDeleteResumes==null||userDeleteResumes.isEmpty()){
-                userDeleteResumeSortedNum = 0;
-            }else{
-                userDeleteResumeSortedNum = userDeleteResumes.getFirst().getUserDeleteResumeSortedNum();
-
-            }
             UserDeleteResume userDeleteResume = new UserDeleteResume();
             userDeleteResume.setUserDeleteResumeId(userSaveResumeInfoVO.getUserSaveResumeId());
             userDeleteResume.setUserDeleteResumeResumeName(userSaveResumeInfoVO.getUserSaveResumeResumeName());
@@ -109,13 +104,21 @@ public class UserDeleteResumeServiceImpl implements UserDeleteResumeService {
             userDeleteResume.setUserDeleteResumeResumeReactCode(userSaveResumeInfoVO.getUserSaveResumeResumeReactCode());
             userDeleteResume.setUserDeleteResumeCreatedTime(userSaveResumeInfoVO.getUserSaveResumeCreatedTime());
             userDeleteResume.setUserDeleteResumeUpdatedTime(userSaveResumeInfoVO.getUserSaveResumeUpdatedTime());
-            userDeleteResume.setUserDeleteResumeSortedNum(userDeleteResumeSortedNum + 1);
+            Integer userDeleteResumeSortedNum = 0;
+            if (userDeleteResumes==null||userDeleteResumes.isEmpty()){
+                userDeleteResume.setUserDeleteResumeSortedNum(userDeleteResumeSortedNum);
+            }else{
+                userDeleteResumeSortedNum = userDeleteResumes.getFirst().getUserDeleteResumeSortedNum();
+                userDeleteResume.setUserDeleteResumeSortedNum(userDeleteResumeSortedNum + 1);
+
+            }
             userDeleteResume.setUserDeleteResumeUserId(userSaveResumeUserId);
             userDeleteResume.setUserDeleteResumeDeleteTime(new Date());
             userDeleteResumeMapper.insert(userDeleteResume);
+        }catch (BusException e){
+            throw e;
         }catch (Exception e){
-            log.error("将用户保存的简历添加到用户删除简历失败");
-            e.printStackTrace();
+            log.error("将用户保存的简历添加到用户删除简历失败", e);
             // TODO:一定要回抛,不然事务失效就会有，删除了但是记录失败 按理说两者要同成功或者同失败的
             throw new RuntimeException("将用户保存的简历添加到用户删除简历失败");
         }
@@ -131,15 +134,7 @@ public class UserDeleteResumeServiceImpl implements UserDeleteResumeService {
             queryWrapper.eq(UserSaveResume::getUserSaveResumeUserId, userId);
             queryWrapper.orderByDesc(UserSaveResume::getUserSaveResumeSortedNum);
             List<UserSaveResume> userSaveResumes = userSaveResumeMapper.selectList(queryWrapper);
-            Integer userSaveResumeSortedNum=0;
-            if (userSaveResumes==null||userSaveResumes.isEmpty()){
-                userSaveResumeSortedNum = 0;
-            }else{
-                userSaveResumeSortedNum = userSaveResumes.getFirst().getUserSaveResumeSortedNum();
-            }
-            if (userSaveResumeSortedNum>=4){
-                throw new BusException(UserCodeEnum.USER_SAVE_RESUME_NOT_DAYU_FIVE);
-            }
+
             UserSaveResume userSaveResume = new UserSaveResume();
             userSaveResume.setUserSaveResumeResumeName(userDeleteResumeInfoVO.getUserDeleteResumeResumeName());
             LambdaQueryWrapper<IndustryMap> queryWrapper1 = new LambdaQueryWrapper<>();
@@ -149,19 +144,29 @@ public class UserDeleteResumeServiceImpl implements UserDeleteResumeService {
             userSaveResume.setUserSaveResumeResumeReactCode(userDeleteResumeInfoVO.getUserDeleteResumeResumeReactCode());
             userSaveResume.setUserSaveResumeCreatedTime(userDeleteResumeInfoVO.getUserDeleteResumeCreatedTime());
             userSaveResume.setUserSaveResumeUpdatedTime(userDeleteResumeInfoVO.getUserDeleteResumeUpdatedTime());
-            userSaveResume.setUserSaveResumeSortedNum(userSaveResumeSortedNum + 1);
+            Integer userSaveResumeSortedNum=0;
+            if (userSaveResumes==null||userSaveResumes.isEmpty()){
+                userSaveResume.setUserSaveResumeSortedNum(userSaveResumeSortedNum);
+            }else{
+                userSaveResumeSortedNum = userSaveResumes.getFirst().getUserSaveResumeSortedNum();
+                userSaveResume.setUserSaveResumeSortedNum(userSaveResumeSortedNum + 1);
+            }
+            if (userSaveResumeSortedNum>=4){
+                throw new BusException(UserCodeEnum.USER_SAVE_RESUME_NOT_DAYU_FIVE);
+            }
             userSaveResume.setUserSaveResumeUserId(userId);
             userSaveResumeMapper.insert(userSaveResume);
             LambdaQueryWrapper<UserDeleteResume> lambdaQueryWrapper2 = new LambdaQueryWrapper<>();
-            lambdaQueryWrapper2.eq(UserDeleteResume::getUserDeleteResumeId, userDeleteResumeInfoVO.getUserDeleteResumeUserId());
+            lambdaQueryWrapper2.eq(UserDeleteResume::getUserDeleteResumeUserId, userDeleteResumeInfoVO.getUserDeleteResumeUserId());
             lambdaQueryWrapper2.eq(UserDeleteResume::getUserDeleteResumeSortedNum, userDeleteResumeInfoVO.getUserDeleteResumeSortedNum());
             userDeleteResumeMapper.delete(lambdaQueryWrapper2);
             reorderResumeSortedNum(userDeleteResumeInfoVO.getUserDeleteResumeUserId(), userDeleteResumeInfoVO.getUserDeleteResumeSortedNum());
             // TODO 添加回去以后要删掉
             log.info("将用户删除简历添加到用户保存简历成功");
+        }catch (BusException e){
+            throw e;
         }catch (Exception e){
-            log.error("添加用户删除简历失败");
-            e.printStackTrace();
+            log.error("添加用户删除简历失败", e);
             throw new RuntimeException("添加用户删除简历失败");
         }
 
@@ -178,8 +183,11 @@ public class UserDeleteResumeServiceImpl implements UserDeleteResumeService {
             userDeleteResumeBySystemService.addExpiredUserDeleteResume(userDeleteResumes);
             userDeleteResumeMapper.delete(queryWrapper);
             log.info("删除用户所有删除简历成功");
+        }catch (BusException e){
+            throw e;
         }catch (Exception e){
-            log.error("删除用户所有删除简历失败");
+            log.error("删除用户所有删除简历失败", e);
+            throw new RuntimeException("删除用户所有删除简历失败");
         }
 
     }
@@ -211,9 +219,22 @@ public class UserDeleteResumeServiceImpl implements UserDeleteResumeService {
                         .collect(Collectors.toList()));
             }
 
+        } catch (BusException e){
+            throw e;
         } catch (Exception e){
-            log.error("系统删除简历表添加失败");
+            log.error("系统删除简历表添加失败", e);
+            throw new RuntimeException("系统删除简历表添加失败");
         }
+    }
+
+    @Override
+    public void updateUserDeleteResumeNameByUserIdAndResumeSortedNumAndResumeName(Integer userId, Integer resumeSortedNum, String resumeName) {
+        LambdaQueryWrapper<UserDeleteResume> queryWrapper = new LambdaQueryWrapper<>();
+        queryWrapper.eq(UserDeleteResume::getUserDeleteResumeUserId, userId);
+        queryWrapper.eq(UserDeleteResume::getUserDeleteResumeSortedNum, resumeSortedNum);
+        UserDeleteResume userDeleteResume = userDeleteResumeMapper.selectOne(queryWrapper);
+        userDeleteResume.setUserDeleteResumeResumeName(resumeName);
+        userDeleteResumeMapper.updateById(userDeleteResume);
     }
 
     private void reorderResumeSortedNum(Integer userId, Integer deletedSortedNum) {
@@ -226,7 +247,7 @@ public class UserDeleteResumeServiceImpl implements UserDeleteResumeService {
         if (resumeList != null && !resumeList.isEmpty()) {
             for (UserDeleteResume resume : resumeList) {
                 LambdaUpdateWrapper<UserDeleteResume> updateWrapper = new LambdaUpdateWrapper<>();
-                updateWrapper.eq(UserDeleteResume::getUserDeleteResumeId, resume.getUserDeleteResumeId());
+                updateWrapper.eq(UserDeleteResume::getUserDeleteResumeUserId, resume.getUserDeleteResumeUserId());
                 updateWrapper.eq(UserDeleteResume::getUserDeleteResumeSortedNum, resume.getUserDeleteResumeSortedNum());
                 updateWrapper.set(UserDeleteResume::getUserDeleteResumeSortedNum, resume.getUserDeleteResumeSortedNum() - 1);
                 userDeleteResumeMapper.update(null, updateWrapper);
